+++ DataAccess/DataAccess.Static/PostgreAdapter.hpp
@@ -20,7 +20,7 @@ private:
     inline static std::mutex ms_static_mutex{};
     inline static std::shared_ptr<PostgreAdapter> msp_instance{};
     inline static constexpr std::string_view ms_defaultOptions =
-        "dbname=postgres user=postgres hostaddr=127.0.0.1 port=5432 password=123";
+        "dbname=juniorgram user=postgres hostaddr=127.0.0.1 port=5432";
 
     std::mutex m_query_mutex;
     std::unique_ptr<pqxx::connection> m_connection;
